moved mutex to util
